Poll component done

+++ shared/database/src/commonMain/kotlin/com/sedsoftware/tackle/database/model/CachedConfig.kt
@@ -30,8 +30,8 @@ internal class CachedConfig(
 
     @Serializable
     class CachedPolls(
-        val maxOptions: Long = 0L,
-        val maxCharactersPerOption: Long = 0L,
+        val maxOptions: Int = 0,
+        val maxCharactersPerOption: Int = 0,
         val minExpiration: Long = 0L,
         val maxExpiration: Long = 0L,
     )
