+++ shared/network/src/commonMain/kotlin/com/sedsoftware/tackle/network/response/InstanceResponse.kt
@@ -53,8 +53,8 @@ internal class InstanceConfigurationAttachmentsResponse(
 
 @Serializable
 internal class InstanceConfigurationPollsResponse(
-    @SerialName("max_options") val maxOptions: Long = 0L,
-    @SerialName("max_characters_per_option") val maxCharactersPerOption: Long = 0L,
+    @SerialName("max_options") val maxOptions: Int = 0,
+    @SerialName("max_characters_per_option") val maxCharactersPerOption: Int = 0,
     @SerialName("min_expiration") val minExpiration: Long = 0L,
     @SerialName("max_expiration") val maxExpiration: Long = 0L,
 )
